Refactored details-toggle JS module and guidance CSS to support new definition pattern
