+++ assets/js/components.js
@@ -3,11 +3,10 @@ import 'babel-polyfill';
 import './focus-styles';
 import './dialog';
 import './submitter';
+import './details-toggle.js';
 // Character count component (textarea)
 import '../../components/02-form-elements/textarea/character-limit.js';
-import '../../components/definitions/definitions.js';
 import '../../components/02-form-elements/input/input--password';
 import '../../components/timeout/timeout.js';
 import '../../components/accordion/accordion.js';
-import '../../components/guidance/details-toggle.js';
 import '../../components/04-navigation/nav.js';
