+++ assets/sass/fixed.scss
@@ -1,13 +1,14 @@
 $responsive: false;
 
-@import "partials/helpers/**/*.scss";
+@import "partials/helpers/*.scss";
 @import "partials/vars/*.scss";
+@import "partials/behaviours/*.scss";
 
 $grid-max-width: 800px;
 
 @import "partials/vendor/normalize.scss";
-@import "partials/base/**/*.scss";
-@import "partials/objects/**/*.scss";
+@import "partials/base/*.scss";
+@import "partials/objects/*.scss";
 @import "../../components/**/*.scss";
-@import "partials/utilities/**/*.scss";
+@import "partials/utilities/*.scss";
