+++ assets/sass/partials/behaviours/_details-toggle.scss
@@ -1,0 +1,47 @@
+%js-details-trigger {
+  &::before {
+    $size: 1rem;
+
+    transition: transform 300ms ease-in-out;
+    margin-right: 0.3rem;
+    speak: none;
+    content: "";
+    background: url("#{$static}/img/icons/icons--guidance.svg") no-repeat center center;
+    background-size: auto;
+    text-align: center;
+    width: $size;
+    height: $size;
+    position: relative;
+    left: -0.1rem;
+    vertical-align: middle;
+    display: inline-block;
+  }
+
+  &:focus {
+    outline: none;
+    &::before {
+      background: url("#{$static}/img/icons/icons--guidance-white.svg") no-repeat center center;
+    }
+  }
+}
+
+%js-details-trigger--is-expanded {
+  &::before {
+    transform: rotate(90deg);
+  }
+}
+
+%js-details-body {
+  overflow: hidden;
+  width: 100%;
+  opacity: 0;
+  transition: all 0;
+  max-height: 0;
+}
+
+%js-details-body--is-expanded {
+  max-height: 10000em;
+  height: auto;
+  opacity: 1;
+  transition: opacity 300ms ease-out, max-height 300ms ease-out;
+}
