+++ assets/sass/responsive.scss
@@ -1,9 +1,10 @@
 $responsive: true;
 
-@import "partials/helpers/**/*.scss";
+@import "partials/helpers/*.scss";
 @import "partials/vars/*.scss";
+@import "partials/behaviours/*.scss";
 @import "partials/vendor/normalize.scss";
-@import "partials/base/**/*.scss";
-@import "partials/objects/**/*.scss";
+@import "partials/base/*.scss";
+@import "partials/objects/*.scss";
 @import "../../components/**/*.scss";
-@import "partials/utilities/**/*.scss";
+@import "partials/utilities/*.scss";
