+++ components/definition/definition.config.js
@@ -1,0 +1,12 @@
+module.exports = {
+  title: 'Definition',
+  status: 'wip',
+  context: {
+    definitiionHideText: 'Hide this',
+    definitionText: 'What is a photovoltaic system?',
+    definitionContent: '<p>A typical photovoltaic system employs solar panels, each comprising\n' +
+      'a number of solar cells, which generate electrical power. PV installations\n' +
+      'may be ground-mounted, rooftop mounted or wall mounted. The mount may be\n' +
+      'fixed, or use a solar tracker to follow the sun across the sky.</p>'
+  }
+};
