--- components/definitions/_definitions.scss
@@ -1,102 +1,0 @@
-.definition {
-  display: none;
-  border-top: 1px solid $color-borders;
-  margin-top: 3em;
-
-  .no-js & {
-    display: block;
-  }
-}
-
-.definition__list {
-  margin: 0;
-  padding: 1.5em 0 1em;
-  position: relative;
-}
-
-.definition__title {
-  display: inline-block;
-  padding-left: 1.5em;
-  @include mq(1100px) {
-    padding-left: 0;
-  }
-
-  dfn {
-    font-style: normal;
-  }
-}
-
-.definition__backtotop {
-  $size: 1.5rem;
-
-  background: url("#{$static}/img/icons/icons--definition-arrow.svg") no-repeat center center;
-  background-size: auto;
-  text-align: center;
-  width: $size;
-  height: $size;
-  display: inline-block;
-  text-indent: -999999px;
-  overflow: hidden;
-  vertical-align: top;
-  position: absolute;
-  top: 1.5em;
-  left: -0.25em;
-  @include mq(1100px) {
-    top: 1.5em;
-    left: -2em;
-  }
-  &:focus,
-  &:active {
-    background-color: white;
-  }
-}
-
-.definition__description {
-  margin: 0;
-}
-
-.definition__anchor {
-  border: 0;
-  text-decoration: none;
-  color: $color-links;
-  font-weight: inherit;
-  padding: 0;
-  border-width: 0 0 2px;
-  border-radius: 0;
-  border-style: dashed;
-  margin-right: 0.6em;
-  appearance: none;
-  background-color: transparent;
-  position: relative;
-  display: inline-block;
-  text-align: left;
-  overflow: visible;
-  &:hover {
-    text-decoration: none;
-    color: $color-links-hover;
-    border-color: $color-links-hover;
-  }
-
-  &:focus {
-    color: #fff;
-    background-color: $color-links;
-    outline: none;
-  }
-
-  &::after {
-    content: "";
-    position: absolute;
-    top: 0;
-    right: -0.6em;
-    background: transparent url("#{$static}/img/icons/icons--definition-anchor.svg") no-repeat 50%;
-    background-size: auto;
-    padding: 0;
-    display: inline-block;
-    white-space: nowrap;
-    line-height: 0;
-    vertical-align: super;
-    margin-left: 0;
-    width: 0.6em;
-    height: 0.6em;
-  }
-}
