--- components/definitions/definitions.hbs
@@ -1,36 +1,0 @@
-<div class="page" style="height:100%">
-  <div class="page__content" style="height:100%">
-    <div class="page__container container">
-      <div class="grid grid--reverse">
-        <div class="grid__col col-4@m pull-1@m">
-          <div class="popup" id="popup"></div>
-        </div>
-        <div class="grid__col col-7@m pull-1@m">
-          <div class="{{nav-wrapper-classes}}">
-            <h1 class="question__title saturn">{{{definitioncontent}}}</h1>
-            <div class="definition">
-              {{#each definitions}}
-                <dl class="definition__list">
-                  {{#if title}}
-                    <dt class="definition__title venus">
-                      <dfn class="venus" id="def-{{id}}">{{title}}</dfn></strong>
-                    </dt>
-                  {{/if}}
-                  {{#if description}}
-                    <dd class="definition__description mars" >
-                      {{description}}
-                      <a href="#anchor-{{id}}" class="definition__backtotop">Back to top</a>
-                    </dd>
-                  {{/if}}
-                </dl>
-              {{/each}}
-            </div>
-          </div>
-        </div>
-        <div class="grid__col col-4@m pull-1@m">
-          <div class="popup" id="popup"></div>
-        </div>
-      </div>
-    </div>
-  </div>
-</div>
