+++ package.json
@@ -9,6 +9,7 @@
     "build": "yarn prestart && yarn compile_assets && gulp scripts:lint && gulp fractal:build",
     "compile_assets": "gulp",
     "start": "yarn && gulp dev",
+    "test": "gulp scripts:test",
     "deploy": "gulp && yarn build && gh-pages -d dist"
   },
   "dependencies": {
